Swap to Typescript file and changed export
